Split media library record helpers
